Related Products Components Update (#31)
* restructured files

* Updated RelateAndCompare.js for sampleStore

* updated RelatedProductsEntry for sample store

* Updated RelatedProducts for sample store

* Updated YourOutfit for sample store
