+++ client/src/components/related-items/comparisonModal/ComparisonModal.js
