+++ client/src/components/related-items/comparisonModal/Features.js
